Separate nix-shell scripts from scripts themselves.
Nix-shell will confuse the nvim lsp and add the nix langauge server.
It doesn't look simple to get around but that's ok because splitting the
files from the nix-shell scripts gives the benefit of having those
scripts be lintable. It also separates the scripts from it's
dependencies if someone else wants to copy them.